requirements.txt
paho-mqtt
aniso8601==8.0.0
antiorm==1.2.1
Click==7.0
db==0.1.1
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.11.1
jsonify==0.5
MarkupSafe==1.1.1
PyJWT==1.7.1
pytz==2019.3
redis==3.4.1
six==1.14.0
Werkzeug==0.16.1
websockets
flask_socketio
opencv-python==4.0.0.21
requests
eventlet
datetime
flask_cors